pgdb/requirements.txt
# PostgreSQL Database Dependencies
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
